fix github_pr endpoint
